refactor: move event handlers to own folders
