+++ src/eventHandlers/pullRequest/index.ts
@@ -1,10 +1,10 @@
 import { info, warning } from '@actions/core';
 import { context, GitHub } from '@actions/github';
 
-import { DEPENDABOT_GITHUB_LOGIN } from '../constants';
-import { approveAndMergePullRequestMutation } from '../graphql/mutations';
+import { DEPENDABOT_GITHUB_LOGIN } from '../../constants';
+import { approveAndMergePullRequestMutation } from '../../graphql/mutations';
 
-export const handle = async (octokit: GitHub): Promise<void> => {
+export const pullRequestHandle = async (octokit: GitHub): Promise<void> => {
   const pullRequest = context.payload.pull_request;
 
   if (pullRequest === undefined) {
